fix: 인터페이스 분리 및 패딩 1.6rem으로 변경
