+++ src/components/order/OrderDeliveryCard/OrderDeliveryCard.module.scss
@@ -1,5 +1,5 @@
 .delivery {
-  margin: 1.2rem 3rem 2rem;
+  margin: 1.2rem 1.6rem 2rem;
 }
 
 .titleAndButton {
@@ -67,7 +67,6 @@
 
 .noDeliveryDescription {
   color: $color-gray-400;
-
   font-size: 1.2rem;
   font-weight: 700;
   line-height: 1.6rem;
