+++ src/components/payment/PaymentAgree.module.scss
@@ -1,5 +1,5 @@
 .agree {
-  margin: 0 3rem;
+  margin: 0 2rem;
 }
 .checkboxTitle {
   display: flex;
