+++ src/components/payment/TotalPay.module.scss
@@ -1,5 +1,5 @@
 .calculateContainer {
-  padding: 0 3rem;
+  padding: 0 1.6rem;
 }
 
 .totalNumberTitle {
