+++ src/components/payment/TotalPay.tsx
@@ -9,6 +9,9 @@ interface TotalPayProps {
 
 export default function TotalPay({ totalPrice, totalOriginalPrice, title, productCount }: TotalPayProps) {
   const discountAmount = totalOriginalPrice - totalPrice;
+  const formattedTotalOriginalPrice = totalOriginalPrice.toLocaleString('ko-KR');
+  const formattedDiscountAmount = discountAmount.toLocaleString('ko-KR');
+  const formattedTotalPrice = totalPrice.toLocaleString('ko-KR');
 
   return (
     <>
@@ -18,15 +21,15 @@ export default function TotalPay({ totalPrice, totalOriginalPrice, title, produc
         <div className={styles.individualCost}>
           <div className={`${styles.pricePair} ${styles.gray}`}>
             <div>원가</div>
-            <div>{totalOriginalPrice}원</div>
+            <div>{formattedTotalOriginalPrice}원</div>
           </div>
           <div className={styles.pricePair}>
-            <div>할인가</div>
-            <div>{totalOriginalPrice - discountAmount}원</div>
+            <div>할인 금액</div>
+            <div>-{formattedDiscountAmount}원</div>
           </div>
           <div className={`${styles.pricePair} ${styles.pink}`}>
-            <div>할인 금액</div>
-            <div>-{discountAmount}원</div>
+            <div>결제 금액</div>
+            <div>{formattedTotalPrice}원</div>
           </div>
           <div className={styles.pricePair}>
             <div>배송비</div>
@@ -36,7 +39,7 @@ export default function TotalPay({ totalPrice, totalOriginalPrice, title, produc
         <div className={styles.line}></div>
         <div className={styles.totalPrice}>
           <div>총 결제 금액</div>
-          <div>{totalPrice}원</div>
+          <div>{formattedTotalPrice}원</div>
         </div>
       </div>
     </>
