+++ src/pages/payment/Payment.module.scss
@@ -1,5 +1,5 @@
 .headerRoot {
-  padding: 2rem 3rem 0 3rem;
+  padding: 2rem 1.6rem 0 1.6rem;
 }
 
 .headerName {
@@ -12,11 +12,11 @@
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  margin: 0 3rem 1.2rem;
+  margin: 0 1.6rem 1.2rem;
 }
 
 .orderProduct {
-  padding: 2rem 3rem 0 3rem;
+  padding: 2rem 1.6rem 0 1.6rem;
 }
 
 .orderTitle {
@@ -45,7 +45,7 @@
 
 .paymentButton {
   margin-top: 3.2rem;
-  padding: 0 3rem;
+  padding: 0 1.6rem;
 }
 
 .paymentAgree {
